requirements.txt
fastapi
uvicorn
pandas
python-multipart
tenseal
azure-storage-blob
numpy